Simplify React context usage
