+++ assets/src/web-stories-block/block/edit.js
@@ -29,7 +29,6 @@ import { __ } from '@wordpress/i18n';
  */
 import { ReactComponent as BlockIcon } from '../images/icon.svg';
 import StoryEmbedEdit from './block-types/story-embed-block/edit';
-import { ConfigProvider } from './config';
 import StoriesBlockControls from './components/storiesBlockControls';
 import BlockConfigurationPanel from './components/storiesBlockConfigurationPanel';
 import LatestStoriesEdit from './block-types/latest-stories/edit';
@@ -41,9 +40,6 @@ import {
   BLOCK_TYPES,
   VIEW_TYPES,
 } from './constants';
-import { webStoriesBlockSettings } from './globals';
-
-const { config } = webStoriesBlockSettings;
 
 function WebStoriesEdit({ attributes, setAttributes, className, isSelected }) {
   const { blockType, viewType } = attributes;
@@ -78,7 +74,7 @@ function WebStoriesEdit({ attributes, setAttributes, className, isSelected }) {
   }
 
   return (
-    <ConfigProvider config={config}>
+    <>
       <StoriesBlockControls
         blockType={blockType}
         viewType={viewType}
@@ -110,7 +106,7 @@ function WebStoriesEdit({ attributes, setAttributes, className, isSelected }) {
           isSelected={isSelected}
         />
       )}
-    </ConfigProvider>
+    </>
   );
 }
